move inference_mode to tree_utils
