+++ src/levanter/main/eval_lm.py
@@ -21,7 +21,7 @@ from levanter.data.text import CausalLmDataset, LMDatasetConfig
 from levanter.models.gpt2 import Gpt2Config
 from levanter.models.lm_model import LmConfig, LmExample, LmHeadModel
 from levanter.trainer import TrainerConfig
-from levanter.utils.jax_utils import inference_mode
+from levanter.utils.tree_utils import inference_mode
 
 
 logger = logging.getLogger(__name__)
