+++ tests/test_levanter_hf_consistency.py
@@ -11,7 +11,7 @@ from levanter.checkpoint import load_checkpoint
 from levanter.models.backpack import BackpackLMHeadModel
 from levanter.models.gpt2 import Gpt2LMHeadModel
 from levanter.trainer import TrainerConfig
-from levanter.utils.jax_utils import inference_mode
+from levanter.utils.tree_utils import inference_mode
 from test_utils import skip_if_checkpoint_not_accessible, skip_if_hf_model_not_accessible, skip_if_no_torch
