+++ tests/test_mpt.py
@@ -8,7 +8,7 @@ from transformers import AutoModelForCausalLM
 import haliax
 
 from levanter.models.mpt import MptConfig, MptLmHeadModel
-from levanter.utils.jax_utils import inference_mode
+from levanter.utils.tree_utils import inference_mode
 from test_utils import check_load_config, parameterize_with_configs, skip_if_no_torch
